Fix 10 bugs from comprehensive script review
CRITICAL: prepare_liver_data.py source/target were swapped (would train
on image as mask and mask as target). Fixed + added DSAD naming support
(--dsad flag) and pre-split directory support (auto-detected).

MEDIUM fixes:
- cldm/model.py: removed broken .to(float16) on safetensors dict
- create_liver_sample.py: json.dump -> JSONL format consistency
- tutorial_inference_local.py: now saves dual-control output too
- create_sample_data.py: added missing os.chdir

LOW fixes:
- tutorial_train_single_gpu.py: STRICT_LOAD param, SD_LOCKED docs
- tutorial_dataset.py: transform to __init__, removed unused cv2
- tutorial_dataset_sample.py: same as above
- SLURM scripts: added --time limits, trap cleanup on failure

+++ cldm/model.py
@@ -24,7 +24,7 @@ def load_state_dict(ckpt_path, location='cpu', exclude_buffers=None):
     _, extension = os.path.splitext(ckpt_path)
     if extension.lower() == ".safetensors":
         import safetensors.torch
-        state_dict = safetensors.torch.load_file(ckpt_path, device=location).to(torch.float16)
+        state_dict = safetensors.torch.load_file(ckpt_path, device=location)
     else:
         # weights_only=False needed: older .ckpt files (SD v1.5, ADC) may contain
         # non-tensor Python objects (e.g. PL callbacks). Safe for trusted checkpoints.
